backend/pyproject.toml
[project]
name = "ai_org_backend"
version = "0.1.0"
description = "AI‑Org backend"
authors = [{ name = "Your Org" }]
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "sqlmodel",
  "celery",
  "redis",
  "neo4j",
  "prometheus-client",
  "pydantic-settings",
  "python-dotenv",
  "jinja2",
  "networkx",
  "qdrant-client",
  "openai==0.27.0",
  "PyJWT",
  "python-multipart",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "black", "alembic"]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
